First commit
Need to fix decimal behavior and touch up how up and down functionality works when the user wants to toggle flow of control (i.e. changing N -> S or E - W on the latitude and longitude examples, respectively. Allow user to type from left to right, implicitly highlighting the editing index based on cursorPositionChanged and moving the cursor to the next available, editable index, if possible. Backspace should clear the index and zero out the highlighted index. The user can only paste from their clipboard if it overlays perfectly with the rules enforced via the Range types. At a high level this is like a pseudo RegEx state machine-like thing.